+++ src/Guilded/client/AbstractGuildedClient.Client.cs
@@ -172,15 +172,16 @@ public abstract partial class AbstractGuildedClient : BaseGuildedClient
 
         // They aren't disposed by DisconnectAsync, only shut down
         Websocket.Dispose();
+        GC.SuppressFinalize(this);
     }
 
-    private void EnforceLimit(string name, string value, short limit)
+    private static void EnforceLimit(string name, string value, short limit)
     {
         if (value.Length > limit)
             throw new ArgumentOutOfRangeException(name, value, $"{name} exceeds the {limit} character limit");
     }
 
-    private void EnforceLimitOnNullable(string name, string? value, short limit)
+    private static void EnforceLimitOnNullable(string name, string? value, short limit)
     {
         if (value is not null) EnforceLimit(name, value, limit);
     }